Move aidl files in package it.bonificamarche.services.aidl
