--- app/src/main/aidl/it/bonificamarche/services/Photo.aidl
@@ -1,4 +1,0 @@
-// Photo.aidl
-package it.bonificamarche.services;
-
-parcelable Photo;
