+++ app/src/main/aidl/it/bonificamarche/services/aidl/Action.aidl
@@ -1,0 +1,4 @@
+// Action.aidl
+package it.bonificamarche.services.aidl;
+
+parcelable Action;
