+++ app/src/main/aidl/it/bonificamarche/services/aidl/Photo.aidl
@@ -1,0 +1,4 @@
+// Photo.aidl
+package it.bonificamarche.services.aidl;
+
+parcelable Photo;
