+++ app/src/main/aidl/it/bonificamarche/services/aidl/Transmission.aidl
@@ -1,4 +1,4 @@
 // Transmission.aidl
-package it.bonificamarche.services;
+package it.bonificamarche.services.aidl;
 
 parcelable Transmission;
